+++ .github/workflows/ci.yaml
@@ -1,0 +1,102 @@
+name: CI
+on: [push, pull_request]
+
+jobs:
+    test-linux:
+        name: Run tests on Linux
+        runs-on: ubuntu-latest
+        timeout-minutes: 20
+        steps:
+            - name: Install atftp
+              run: |
+                  sudo apt-get update
+                  sudo apt-get install -y atftp
+            - name: Install beta toolchain
+              uses: actions-rs/toolchain@v1
+              with:
+                  profile: minimal
+                  toolchain: beta
+                  override: true
+            - uses: actions/checkout@v1
+            - name: Run tests
+              uses: actions-rs/cargo@v1
+              with:
+                  command: test
+                  args: --release --features external-client-tests
+                  toolchain: beta
+            - name: Run long tests
+              uses: actions-rs/cargo@v1
+              with:
+                  command: test
+                  args: --release --features external-client-tests -- --ignored
+                  toolchain: beta
+    test-windows:
+        name: Run tests on Windows
+        runs-on: windows-latest
+        timeout-minutes: 20
+        steps:
+            - name: Install tftp
+              shell: pwsh
+              run: |
+                  Invoke-WebRequest -OutFile ${env:TEMP}\tftp.exe https://www.winagents.com/downloads/tftp.exe
+            - name: Install beta toolchain
+              uses: actions-rs/toolchain@v1
+              with:
+                  profile: minimal
+                  toolchain: beta
+                  override: true
+            - uses: actions/checkout@v1
+            - name: Run tests
+              uses: actions-rs/cargo@v1
+              with:
+                  command: test
+                  args: --release --features external-client-tests
+                  toolchain: beta
+            - name: Run long tests
+              uses: actions-rs/cargo@v1
+              with:
+                  command: test
+                  args: --release --features external-client-tests -- --ignored
+                  toolchain: beta
+    test-macos:
+        name: Run tests on macOS
+        runs-on: macOS-latest
+        timeout-minutes: 20
+        steps:
+            - name: Install beta toolchain
+              uses: actions-rs/toolchain@v1
+              with:
+                  profile: minimal
+                  toolchain: beta
+                  override: true
+            - uses: actions/checkout@v1
+            - name: Run tests
+              uses: actions-rs/cargo@v1
+              with:
+                  command: test
+                  args: --release
+                  toolchain: beta
+    lints:
+        name: Lints
+        runs-on: ubuntu-latest
+        steps:
+            - name: Install beta toolchain
+              uses: actions-rs/toolchain@v1
+              with:
+                  profile: minimal
+                  toolchain: beta
+                  override: true
+                  components: rustfmt, clippy
+            - uses: actions/checkout@v1
+            - name: Run cargo fmt
+              uses: actions-rs/cargo@v1
+              with:
+                  command: fmt
+                  args: --all -- --check
+                  toolchain: beta
+            - name: Run cargo clippy
+              uses: actions-rs/cargo@v1
+              with:
+                  command: clippy
+                  args: -- -D clippy::all
+                  toolchain: beta
