--- .travis.yml
@@ -1,3 +1,0 @@
-language: rust
-rust:
-    - nightly
