+++ Cargo.toml
@@ -31,4 +31,4 @@ matches = "0.1.8"
 tracing-subscriber = "0.1.5"
 
 [features]
-atftp-tests = []
+external-client-tests = []
