+++ src/tests/mod.rs
@@ -1,8 +1,7 @@
 #![cfg(test)]
 
+mod external_client;
 mod handles;
 mod packet;
 mod random_file;
-
-#[cfg(feature = "atftp-tests")]
-mod atftp_rrq;
+mod rrq;
