+++ src/tests/rrq.rs
@@ -1,33 +1,10 @@
-use async_std::task;
-use std::fs;
-use std::io;
-use std::net::SocketAddr;
-use std::process::Command;
-use tempfile::tempdir;
+#![cfg(feature = "external-client-tests")]
 
+use async_std::task;
+
+use super::external_client::*;
 use super::handles::*;
 use crate::AsyncTftpServer;
-
-pub fn atftp_recv(
-    filename: &str,
-    server: SocketAddr,
-) -> io::Result<md5::Digest> {
-    let tmp = tempdir()?;
-    let path = tmp.path().join("data");
-
-    Command::new("atftp")
-        .arg("-g")
-        .arg("-l")
-        .arg(&path)
-        .arg("-r")
-        .arg(filename)
-        .arg(server.ip().to_string())
-        .arg(server.port().to_string())
-        .status()?;
-
-    let data = fs::read(path)?;
-    Ok(md5::compute(data))
-}
 
 fn transfer(file_size: usize) {
     task::block_on(async {
@@ -38,14 +15,15 @@ fn transfer(file_size: usize) {
         let tftpd = AsyncTftpServer::bind(handle, "127.0.0.1:0").await.unwrap();
         let addr = tftpd.local_addr().unwrap();
 
-        // start atftp client
-        let atftp_recv = task::spawn_blocking(move || atftp_recv("test", addr));
+        // start client
+        let tftp_recv =
+            task::spawn_blocking(move || external_tftp_recv("test", addr));
 
         // start server
         tftpd.serve().await.unwrap();
 
         // check md5
-        let recved_md5 = atftp_recv.await.expect("failed to run atftp");
+        let recved_md5 = tftp_recv.await.expect("failed to run tftp client");
         let sent_md5 = md5.lock().unwrap().expect("no md5");
         assert_eq!(recved_md5, sent_md5);
     });
